auth-service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
motor==3.7.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
